[refactor] WBJ-571 Settings page hooks refactored
