+++ src/pages/settings/CreateTeam.tsx
@@ -1,16 +1,15 @@
 import { Link } from 'react-router-dom';
 import { Row, Col, Card, Button, ProgressBar } from 'react-bootstrap';
 import { Wizard, Steps, Step } from 'react-albus';
-import React, { useState } from 'react';
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import CreateTeamNameForm from './CreateTeamNameForm';
 import CreateTeamImageForm from './CreateTeamImageForm';
 import InviteMemberButton from './InviteMemberButton';
+import { useTeamSettings } from './hooks';
 
 const CreateTeam = () => {
-  const [teamId, setTeamId] = useState<number>();
-  const [teamName, setTeamName] = useState<string>();
+  const { team, fileInput, previewImage, onCreateTeam, onChangeLogoImage, onSaveLogoImage, onCancelChangeLogoImage, onInvite } = useTeamSettings();
 
   return (
     <>
@@ -42,12 +41,12 @@ const CreateTeam = () => {
                       <Step
                         id="create"
                         render={({ next }) => (
-                          <CreateTeamNameForm next={next} setTeamId={setTeamId} setTeamName={setTeamName} />
+                          <CreateTeamNameForm next={next} onCreateTeam={onCreateTeam} />
                         )}
                       />
-                      <Step id="logo" render={({ next }) => <CreateTeamImageForm next={next} teamId={teamId} />} />
+                      <Step id="logo" render={({ next }) => <CreateTeamImageForm next={next} fileInput={fileInput} previewImage={previewImage} onChangeLogoImage={onChangeLogoImage} onSaveLogoImage={onSaveLogoImage} onCancelChangeLogoImage={onCancelChangeLogoImage} />} />
                       <Step
-                        id="dumbledore"
+                        id="finished"
                         render={() => (
                           <Row>
                             <Col sm={12}>
@@ -55,12 +54,12 @@ const CreateTeam = () => {
                                 <h2 className="mt-0">
                                   <i className="mdi mdi-check-all"></i>
                                 </h2>
-                                <h3 className="mt-0 mb-4">Team {teamName} Created !</h3>
+                                <h3 className="mt-0 mb-4">Team {team?.name} Created !</h3>
                               </div>
                             </Col>
 
                             <Col className="d-flex justify-content-between">
-                              <InviteMemberButton teamId={teamId} />
+                              <InviteMemberButton onInvite={onInvite} />
                               <Link to="/settings/team">
                                 <Button variant="primary">Done</Button>
                               </Link>
