+++ src/pages/settings/PersonalImageForm.tsx
@@ -1,61 +1,8 @@
-import React, { useState, useRef } from 'react';
 import { Button } from 'react-bootstrap';
-import { toast } from 'react-toastify';
-import 'react-toastify/dist/ReactToastify.css';
-import { createProfileImage as createProfileImageAPI, uploadProfileImage as uploadProfileImageAPI } from 'helpers';
-import { APICore } from 'helpers/api/apiCore';
+import { usePersonalSettings } from './hooks';
 
 const PersonalImageForm: React.FC = () => {
-  const api = new APICore();
-  const user = api.getLoggedInUser();
-  const [previewImage, setPreviewImage] = useState<string>(user.profileImageUrl);
-  const [profileImageFile, setProfileImageFile] = useState<null | File>();
-
-  const fileInput = useRef<HTMLInputElement>(null);
-
-  const onChange = (e: React.ChangeEvent<HTMLInputElement>): void => {
-    if (!e.target.files) {
-      setPreviewImage(user.profileImageUrl);
-      return;
-    } else {
-      setProfileImageFile(e.target.files[0]);
-    }
-
-    const reader = new FileReader();
-
-    reader.onload = () => {
-      if (reader.readyState === 2) {
-        setPreviewImage(reader.result as string);
-      }
-    };
-
-    reader.readAsDataURL(e.target.files[0]);
-  };
-
-  const onCancelProfileImageChange = () => {
-    setPreviewImage(user.profileImageUrl);
-    setProfileImageFile(null);
-  };
-
-  const onSaveProfileImage = async (e: React.MouseEvent<HTMLElement>) => {
-    if (!profileImageFile) {
-      return;
-    } else {
-      const createProfileImageAPIResponse = await createProfileImageAPI();
-      const presignedURL = createProfileImageAPIResponse.data.preSignedURL;
-      await uploadProfileImageAPI(presignedURL, profileImageFile);
-      notify();
-    }
-  };
-
-  const notify = () =>
-    toast(
-      <div>
-        Profile image saved successfully!
-        <br />
-        The change might take a few minutes to be applied.
-      </div>
-    );
+  const { previewImage, fileInput, onChangeProfileImage, onSaveProfileImage, onCancelChangeProfileImage } = usePersonalSettings();
 
   return (
     <>
@@ -77,7 +24,7 @@ const PersonalImageForm: React.FC = () => {
             accept="image/png"
             style={{ display: 'none' }}
             name="profileImage"
-            onChange={onChange}
+            onChange={onChangeProfileImage}
             ref={fileInput}
           />
           <div
@@ -95,7 +42,7 @@ const PersonalImageForm: React.FC = () => {
       <Button className="me-2" onClick={onSaveProfileImage}>
         Save
       </Button>
-      <Button className="btn btn-secondary" onClick={onCancelProfileImageChange}>
+      <Button className="btn btn-secondary" onClick={onCancelChangeProfileImage}>
         Cancel
       </Button>
     </>
