+++ src/pages/settings/TeamInfoForm.tsx
@@ -1,25 +1,8 @@
 import { Button } from 'react-bootstrap';
 import { FormInput } from 'components';
-import { toast } from 'react-toastify';
-import 'react-toastify/dist/ReactToastify.css';
 import { TeamInfoFormProps } from './types';
-import { updateTeamInfo as updateTeamInfoAPI } from 'helpers';
 
-const TeamInfoForm: React.FC<TeamInfoFormProps> = ({ team, teamname, setTeamname }: TeamInfoFormProps) => {
-  const onSubmitTeamInfo = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    if (!team) {
-      return;
-    }
-    const name = ((e.target as HTMLFormElement).elements as { [key: string]: any })['name'].value;
-    await updateTeamInfoAPI(team.id, { name });
-    setTeamname(name);
-    (e.target as HTMLFormElement).reset();
-    changeInfoNotify();
-  };
-
-  const changeInfoNotify = () => toast(<div>Team Information changed successfully!</div>);
-
+const TeamInfoForm: React.FC<TeamInfoFormProps> = ({ teamname, onSubmitTeamInfo }: TeamInfoFormProps) => {
   return (
     <form onSubmit={onSubmitTeamInfo}>
       <FormInput
