+++ src/pages/settings/hooks/index.ts
@@ -1,0 +1,2 @@
+export { default as usePersonalSettings } from './usePersonalSettings';
+export { default as useTeamSettings } from './useTeamSettings';
